kimono_ai_customer_service/requirements.txt
# Kimono AI Customer Service - Dependencies
# Python 3.9+ required

# Core AI/ML
langchain>=0.1.0
langchain-community>=0.0.10
dashscope>=1.14.0  # Alibaba Qwen API
openai>=1.0.0  # For embeddings

# Vector Database
pinecone>=5.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0  # Excel file support
beautifulsoup4>=4.12.0  # HTML parsing
lxml>=4.9.0  # HTML parser backend

# Language Detection
fasttext-langdetect>=1.0.5
langdetect>=1.0.9

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0  # Async SQLite support
asyncpg>=0.29.0  # Async PostgreSQL (for production)

# Authentication
python-jose[cryptography]>=3.3.0  # JWT
passlib[bcrypt]>=1.7.4  # Password hashing
bcrypt>=4.0.0  # bcrypt backend

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# CLI
rich>=13.0.0  # Beautiful terminal output
click>=8.1.0  # CLI framework
